admin/requirements.txt
Flask==3.0.0
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
prometheus-flask-exporter==0.23.0
pytest==8.0.0
pytest-flask==1.3.0
pytest-cov==4.1.0
